refactor back to what i want to merge
