+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/proxy/cdk-requests-go-through-a-proxy-when-configured.integtest.ts
@@ -28,22 +28,8 @@ integTest('requests go through a proxy when configured',
         },
       });
 
-      const connections = JSON.parse(await fs.readFile(path.join(cdkCacheDir, 'connection.json'), 'utf8'));
-      // eslint-disable-next-line no-console
-      console.log('connections', connections);
-
       const requests = await proxyServer.getSeenRequests();
-      const urls = requests.map(req => req.url);
-      // eslint-disable-next-line no-console
-      console.log('1', JSON.stringify(urls));
-      // eslint-disable-next-line no-console
-      console.log('2', JSON.stringify(urls.reverse()));
-
-      const urls2 = urls.filter(u => u.startsWith('https://cli.cdk.dev'));
-      // eslint-disable-next-line no-console
-      console.log('3', urls2);
-
-      expect(urls)
+      expect(requests.map(req => req.url))
         .toContain('https://cli.cdk.dev-tools.aws.dev/notices.json');
 
       const actionsUsed = awsActionsFromRequests(requests);
