+++ packages/@aws-cdk/toolkit-lib/lib/api/network-detector/index.ts
@@ -1,0 +1,1 @@
+export * from './network-detector';
