+++ packages/@aws-cdk/toolkit-lib/lib/api/network-detector/network-detector.ts
@@ -2,7 +2,7 @@ import * as https from 'node:https';
 import type { RequestOptions } from 'node:https';
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import { cdkCacheDir } from './';
+import { cdkCacheDir } from '../../util';
 
 interface CachedConnectivity {
   expiration: number;
@@ -46,7 +46,7 @@ export class NetworkDetector {
     }
   }
 
-  private static readonly TIMEOUT_MS = 500;
+  // private static readonly TIMEOUT_MS = 500;
   private static readonly URL = 'https://cli.cdk.dev-tools.aws.dev/notices.json';
 
   private static async load(): Promise<CachedConnectivity> {
@@ -77,7 +77,7 @@ export class NetworkDetector {
     const options: RequestOptions = {
       // method: 'HEAD',
       agent: agent,
-      timeout: this.TIMEOUT_MS,
+      // timeout: this.TIMEOUT_MS,
     };
 
     return new Promise((resolve) => {
