+++ packages/@aws-cdk/toolkit-lib/lib/index.ts
@@ -17,6 +17,3 @@ export * from './api/cloud-assembly';
 export * from './api/io';
 export * from './api/tags';
 export * from './api/plugin';
-
-// Utilities
-export { NetworkDetector } from './util/network-detector';
