+++ packages/aws-cdk/lib/api/network-detector.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/no-relative-packages */
+export * from '../../../@aws-cdk/toolkit-lib/lib/api/network-detector';
