+++ packages/aws-cdk/lib/cli/telemetry/sink/endpoint-sink.ts
@@ -2,7 +2,8 @@ import type { IncomingMessage } from 'http';
 import type { Agent } from 'https';
 import { request } from 'https';
 import { parse, type UrlWithStringQuery } from 'url';
-import { ToolkitError, NetworkDetector } from '@aws-cdk/toolkit-lib';
+import { ToolkitError } from '@aws-cdk/toolkit-lib';
+import { NetworkDetector } from '../../../api/network-detector';
 import { IoHelper } from '../../../api-private';
 import type { IIoHost } from '../../io-host';
 import type { TelemetrySchema } from '../schema';
