+++ packages/aws-cdk/test/cli/telemetry/sink/funnel.test.ts
@@ -1,9 +1,9 @@
 import * as https from 'https';
 import * as os from 'os';
 import * as path from 'path';
-import { NetworkDetector } from '@aws-cdk/toolkit-lib';
 import * as fs from 'fs-extra';
 import { createTestEvent } from './util';
+import { NetworkDetector } from '../../../../lib/api/network-detector';
 import { IoHelper } from '../../../../lib/api-private';
 import { CliIoHost } from '../../../../lib/cli/io-host';
 import { EndpointTelemetrySink } from '../../../../lib/cli/telemetry/sink/endpoint-sink';
@@ -16,8 +16,7 @@ jest.mock('https', () => ({
 }));
 
 // Mock NetworkDetector
-jest.mock('@aws-cdk/toolkit-lib', () => ({
-  ...jest.requireActual('@aws-cdk/toolkit-lib'),
+jest.mock('../../../../lib/api/network-detector', () => ({
   NetworkDetector: {
     hasConnectivity: jest.fn(),
   },
